GEODE-3643: Add function execution on specific member
fixups from review comments
