+++ geode-core/src/main/java/org/apache/geode/distributed/internal/ClusterDistributionManager.java
@@ -27,6 +27,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.NoSuchElementException;
+import java.util.Objects;
 import java.util.Set;
 import java.util.concurrent.BlockingQueue;
 import java.util.concurrent.ConcurrentHashMap;
@@ -1454,11 +1455,11 @@ public class ClusterDistributionManager implements DistributionManager {
   @Override
   public DistributedMember getMemberWithName(String name) {
     for (DistributedMember id : members.values()) {
-      if (id.getName() != null && id.getName().equals(name)) {
+      if (Objects.equals(id.getName(), name)) {
         return id;
       }
     }
-    if (localAddress.getName() != null && localAddress.getName().equals(name)) {
+    if (Objects.equals(localAddress, name)) {
       return localAddress;
     }
     return null;
