+++ geode-protobuf/src/test/java/org/apache/geode/internal/protocol/protobuf/v1/ExecuteFunctionOnMemberIntegrationTest.java
@@ -26,6 +26,7 @@ import static org.mockito.Mockito.when;
 
 import java.io.IOException;
 import java.net.Socket;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
@@ -61,10 +62,11 @@ import org.apache.geode.security.SecurityManager;
 import org.apache.geode.test.junit.categories.IntegrationTest;
 
 @Category(IntegrationTest.class)
-public class ExecutFunctionOnRegionIntegrationTest {
+public class ExecuteFunctionOnMemberIntegrationTest {
   private static final String TEST_REGION = "testRegion";
   private static final String TEST_FUNCTION_ID = "testFunction";
-  public static final String SECURITY_PRINCIPAL = "principle";
+  private static final String SECURITY_PRINCIPAL = "principle";
+  private static final String SERVER_NAME = "pericles";
   private ProtobufSerializationService serializationService;
   private Socket socket;
   private Cache cache;
@@ -79,6 +81,7 @@ public class ExecutFunctionOnRegionIntegrationTest {
     cacheFactory.set(ConfigurationProperties.MCAST_PORT, "0");
     cacheFactory.set(ConfigurationProperties.ENABLE_CLUSTER_CONFIGURATION, "false");
     cacheFactory.set(ConfigurationProperties.USE_CLUSTER_CONFIGURATION, "false");
+    cacheFactory.set(ConfigurationProperties.NAME, SERVER_NAME);
 
     securityManager = mock(SecurityManager.class);
     cacheFactory.setSecurityManager(securityManager);
@@ -94,7 +97,7 @@ public class ExecutFunctionOnRegionIntegrationTest {
 
     RegionFactory<Object, Object> regionFactory = cache.createRegionFactory();
     regionFactory.setDataPolicy(DataPolicy.PARTITION);
-    Region<Object, Object> testRegion = regionFactory.create(TEST_REGION);
+    regionFactory.create(TEST_REGION);
 
 
     System.setProperty("geode.feature-protobuf-protocol", "true");
@@ -173,12 +176,12 @@ public class ExecutFunctionOnRegionIntegrationTest {
     assertEquals(ClientProtocol.Message.MessageTypeCase.RESPONSE,
         responseMessage.getMessageTypeCase());
     final ClientProtocol.Response response = responseMessage.getResponse();
-    assertEquals(ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONREGIONRESPONSE,
+    assertEquals(ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONMEMBERRESPONSE,
         response.getResponseAPICase());
-    final FunctionAPI.ExecuteFunctionOnRegionResponse executeFunctionOnRegionResponse =
-        response.getExecuteFunctionOnRegionResponse();
+    final FunctionAPI.ExecuteFunctionOnMemberResponse executeFunctionOnMemberResponse =
+        response.getExecuteFunctionOnMemberResponse();
 
-    assertEquals(0, executeFunctionOnRegionResponse.getResultsCount());
+    assertEquals(0, executeFunctionOnMemberResponse.getResultsCount());
 
     Awaitility.await().atMost(5, TimeUnit.SECONDS).until(() -> testFunction.getContext() != null);
   }
@@ -190,13 +193,13 @@ public class ExecutFunctionOnRegionIntegrationTest {
     FunctionService.registerFunction(testFunction);
     final ClientProtocol.Message responseMessage = authenticateAndSendMessage();
 
-    final FunctionAPI.ExecuteFunctionOnRegionResponse executeFunctionOnRegionResponse =
+    final FunctionAPI.ExecuteFunctionOnMemberResponse executeFunctionOnMemberResponse =
         getFunctionResponse(responseMessage);
 
-    assertEquals(1, executeFunctionOnRegionResponse.getResultsCount());
+    assertEquals(1, executeFunctionOnMemberResponse.getResultsCount());
 
     final Object responseValue =
-        serializationService.decode(executeFunctionOnRegionResponse.getResults(0));
+        serializationService.decode(executeFunctionOnMemberResponse.getResults(0));
     assertTrue(responseValue instanceof Integer);
     assertEquals(22, responseValue);
   }
@@ -242,13 +245,13 @@ public class ExecutFunctionOnRegionIntegrationTest {
     FunctionService.registerFunction(testFunction);
     final ClientProtocol.Message responseMessage = authenticateAndSendMessage();
 
-    final FunctionAPI.ExecuteFunctionOnRegionResponse executeFunctionOnRegionResponse =
+    final FunctionAPI.ExecuteFunctionOnMemberResponse executeFunctionOnMemberResponse =
         getFunctionResponse(responseMessage);
 
-    assertEquals(1, executeFunctionOnRegionResponse.getResultsCount());
+    assertEquals(1, executeFunctionOnMemberResponse.getResultsCount());
 
     final Object responseValue =
-        serializationService.decode(executeFunctionOnRegionResponse.getResults(0));
+        serializationService.decode(executeFunctionOnMemberResponse.getResults(0));
     assertNull(responseValue);
   }
 
@@ -258,40 +261,15 @@ public class ExecutFunctionOnRegionIntegrationTest {
         new TestFunction<>(functionContext -> functionContext.getArguments(), true);
     FunctionService.registerFunction(testFunction);
     ClientProtocol.Message.Builder message = createRequestMessageBuilder(
-        FunctionAPI.ExecuteFunctionOnRegionRequest.newBuilder().setFunctionID(TEST_FUNCTION_ID)
-            .setRegion(TEST_REGION).setArguments(serializationService.encode("hello")));
+        FunctionAPI.ExecuteFunctionOnMemberRequest.newBuilder().setFunctionID(TEST_FUNCTION_ID)
+            .addMemberName(SERVER_NAME).setArguments(serializationService.encode("hello")));
 
     authenticateWithServer();
     final ClientProtocol.Message responseMessage = writeMessage(message.build());
 
-    FunctionAPI.ExecuteFunctionOnRegionResponse response = getFunctionResponse(responseMessage);
+    FunctionAPI.ExecuteFunctionOnMemberResponse response = getFunctionResponse(responseMessage);
 
     assertEquals("hello", serializationService.decode(response.getResults(0)));
-  }
-
-  @Test
-  public void filterIsPassedToFunction() throws Exception {
-    final TestFunction<Object> testFunction = new TestFunction<>(context -> "result", true);
-    FunctionService.registerFunction(testFunction);
-    Set<Object> expectedFilter = new HashSet<>(Arrays.asList("key1", "key2"));
-
-    final ClientProtocol.Message.Builder message = createRequestMessageBuilder(
-        FunctionAPI.ExecuteFunctionOnRegionRequest.newBuilder().setFunctionID(TEST_FUNCTION_ID)
-            .setRegion(TEST_REGION).addKeyFilter(serializationService.encode("key1"))
-            .addKeyFilter(serializationService.encode("key2")));
-
-    authenticateWithServer();
-    final ClientProtocol.Message responseMessage = writeMessage(message.build());
-
-    FunctionAPI.ExecuteFunctionOnRegionResponse response = getFunctionResponse(responseMessage);
-    assertEquals("result", serializationService.decode(response.getResults(0)));
-
-    final RegionFunctionContext context = (RegionFunctionContext) testFunction.getContext();
-
-    final Set<?> filter = context.getFilter();
-
-    assertEquals(expectedFilter, filter);
-
   }
 
   @Test
@@ -305,10 +283,10 @@ public class ExecutFunctionOnRegionIntegrationTest {
         return Arrays.asList(requiredPermission);
       }
     };
+
     FunctionService.registerFunction(testFunction);
 
     when(securityManager.authenticate(any())).thenReturn(SECURITY_PRINCIPAL);
-
     when(securityManager.authorize(eq(SECURITY_PRINCIPAL), eq(requiredPermission)))
         .thenReturn(false);
 
@@ -319,12 +297,12 @@ public class ExecutFunctionOnRegionIntegrationTest {
     verify(securityManager).authorize(eq(SECURITY_PRINCIPAL), eq(requiredPermission));
   }
 
-  private FunctionAPI.ExecuteFunctionOnRegionResponse getFunctionResponse(
+  private FunctionAPI.ExecuteFunctionOnMemberResponse getFunctionResponse(
       ClientProtocol.Message responseMessage) {
     assertEquals(responseMessage.getResponse().toString(),
-        ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONREGIONRESPONSE,
+        ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONMEMBERRESPONSE,
         responseMessage.getResponse().getResponseAPICase());
-    return responseMessage.getResponse().getExecuteFunctionOnRegionResponse();
+    return responseMessage.getResponse().getExecuteFunctionOnMemberResponse();
   }
 
   private void authenticateWithServer() throws IOException {
@@ -343,17 +321,17 @@ public class ExecutFunctionOnRegionIntegrationTest {
     authenticateWithServer();
 
     final ClientProtocol.Message request =
-        createRequestMessageBuilder(FunctionAPI.ExecuteFunctionOnRegionRequest.newBuilder()
-            .setFunctionID(TEST_FUNCTION_ID).setRegion(TEST_REGION)).build();
+        createRequestMessageBuilder(FunctionAPI.ExecuteFunctionOnMemberRequest.newBuilder()
+            .setFunctionID(TEST_FUNCTION_ID).addMemberName(SERVER_NAME)).build();
 
     return writeMessage(request);
   }
 
 
   private ClientProtocol.Message.Builder createRequestMessageBuilder(
-      FunctionAPI.ExecuteFunctionOnRegionRequest.Builder functionRequest) {
+      FunctionAPI.ExecuteFunctionOnMemberRequest.Builder functionRequest) {
     return ClientProtocol.Message.newBuilder().setRequest(
-        ClientProtocol.Request.newBuilder().setExecuteFunctionOnRegionRequest(functionRequest));
+        ClientProtocol.Request.newBuilder().setExecuteFunctionOnMemberRequest(functionRequest));
   }
 
   private ClientProtocol.Message writeMessage(ClientProtocol.Message request) throws IOException {
