+++ geode-protobuf/src/test/java/org/apache/geode/internal/protocol/protobuf/v1/ExecuteFunctionOnRegionIntegrationTest.java
@@ -26,7 +26,6 @@ import static org.mockito.Mockito.when;
 
 import java.io.IOException;
 import java.net.Socket;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
@@ -62,11 +61,10 @@ import org.apache.geode.security.SecurityManager;
 import org.apache.geode.test.junit.categories.IntegrationTest;
 
 @Category(IntegrationTest.class)
-public class ExecutFunctionOnMemberIntegrationTest {
+public class ExecuteFunctionOnRegionIntegrationTest {
   private static final String TEST_REGION = "testRegion";
   private static final String TEST_FUNCTION_ID = "testFunction";
-  private static final String SECURITY_PRINCIPAL = "principle";
-  private static final String SERVER_NAME = "pericles";
+  public static final String SECURITY_PRINCIPAL = "principle";
   private ProtobufSerializationService serializationService;
   private Socket socket;
   private Cache cache;
@@ -81,7 +79,6 @@ public class ExecutFunctionOnMemberIntegrationTest {
     cacheFactory.set(ConfigurationProperties.MCAST_PORT, "0");
     cacheFactory.set(ConfigurationProperties.ENABLE_CLUSTER_CONFIGURATION, "false");
     cacheFactory.set(ConfigurationProperties.USE_CLUSTER_CONFIGURATION, "false");
-    cacheFactory.set(ConfigurationProperties.NAME, SERVER_NAME);
 
     securityManager = mock(SecurityManager.class);
     cacheFactory.setSecurityManager(securityManager);
@@ -97,7 +94,7 @@ public class ExecutFunctionOnMemberIntegrationTest {
 
     RegionFactory<Object, Object> regionFactory = cache.createRegionFactory();
     regionFactory.setDataPolicy(DataPolicy.PARTITION);
-    regionFactory.create(TEST_REGION);
+    Region<Object, Object> testRegion = regionFactory.create(TEST_REGION);
 
 
     System.setProperty("geode.feature-protobuf-protocol", "true");
@@ -176,12 +173,12 @@ public class ExecutFunctionOnMemberIntegrationTest {
     assertEquals(ClientProtocol.Message.MessageTypeCase.RESPONSE,
         responseMessage.getMessageTypeCase());
     final ClientProtocol.Response response = responseMessage.getResponse();
-    assertEquals(ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONMEMBERRESPONSE,
+    assertEquals(ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONREGIONRESPONSE,
         response.getResponseAPICase());
-    final FunctionAPI.ExecuteFunctionOnMemberResponse executeFunctionOnMemberResponse =
-        response.getExecuteFunctionOnMemberResponse();
+    final FunctionAPI.ExecuteFunctionOnRegionResponse executeFunctionOnRegionResponse =
+        response.getExecuteFunctionOnRegionResponse();
 
-    assertEquals(0, executeFunctionOnMemberResponse.getResultsCount());
+    assertEquals(0, executeFunctionOnRegionResponse.getResultsCount());
 
     Awaitility.await().atMost(5, TimeUnit.SECONDS).until(() -> testFunction.getContext() != null);
   }
@@ -193,13 +190,13 @@ public class ExecutFunctionOnMemberIntegrationTest {
     FunctionService.registerFunction(testFunction);
     final ClientProtocol.Message responseMessage = authenticateAndSendMessage();
 
-    final FunctionAPI.ExecuteFunctionOnMemberResponse executeFunctionOnMemberResponse =
+    final FunctionAPI.ExecuteFunctionOnRegionResponse executeFunctionOnRegionResponse =
         getFunctionResponse(responseMessage);
 
-    assertEquals(1, executeFunctionOnMemberResponse.getResultsCount());
+    assertEquals(1, executeFunctionOnRegionResponse.getResultsCount());
 
     final Object responseValue =
-        serializationService.decode(executeFunctionOnMemberResponse.getResults(0));
+        serializationService.decode(executeFunctionOnRegionResponse.getResults(0));
     assertTrue(responseValue instanceof Integer);
     assertEquals(22, responseValue);
   }
@@ -245,13 +242,13 @@ public class ExecutFunctionOnMemberIntegrationTest {
     FunctionService.registerFunction(testFunction);
     final ClientProtocol.Message responseMessage = authenticateAndSendMessage();
 
-    final FunctionAPI.ExecuteFunctionOnMemberResponse executeFunctionOnMemberResponse =
+    final FunctionAPI.ExecuteFunctionOnRegionResponse executeFunctionOnRegionResponse =
         getFunctionResponse(responseMessage);
 
-    assertEquals(1, executeFunctionOnMemberResponse.getResultsCount());
+    assertEquals(1, executeFunctionOnRegionResponse.getResultsCount());
 
     final Object responseValue =
-        serializationService.decode(executeFunctionOnMemberResponse.getResults(0));
+        serializationService.decode(executeFunctionOnRegionResponse.getResults(0));
     assertNull(responseValue);
   }
 
@@ -261,15 +258,40 @@ public class ExecutFunctionOnMemberIntegrationTest {
         new TestFunction<>(functionContext -> functionContext.getArguments(), true);
     FunctionService.registerFunction(testFunction);
     ClientProtocol.Message.Builder message = createRequestMessageBuilder(
-        FunctionAPI.ExecuteFunctionOnMemberRequest.newBuilder().setFunctionID(TEST_FUNCTION_ID)
-            .addMemberName(SERVER_NAME).setArguments(serializationService.encode("hello")));
+        FunctionAPI.ExecuteFunctionOnRegionRequest.newBuilder().setFunctionID(TEST_FUNCTION_ID)
+            .setRegion(TEST_REGION).setArguments(serializationService.encode("hello")));
 
     authenticateWithServer();
     final ClientProtocol.Message responseMessage = writeMessage(message.build());
 
-    FunctionAPI.ExecuteFunctionOnMemberResponse response = getFunctionResponse(responseMessage);
+    FunctionAPI.ExecuteFunctionOnRegionResponse response = getFunctionResponse(responseMessage);
 
     assertEquals("hello", serializationService.decode(response.getResults(0)));
+  }
+
+  @Test
+  public void filterIsPassedToFunction() throws Exception {
+    final TestFunction<Object> testFunction = new TestFunction<>(context -> "result", true);
+    FunctionService.registerFunction(testFunction);
+    Set<Object> expectedFilter = new HashSet<>(Arrays.asList("key1", "key2"));
+
+    final ClientProtocol.Message.Builder message = createRequestMessageBuilder(
+        FunctionAPI.ExecuteFunctionOnRegionRequest.newBuilder().setFunctionID(TEST_FUNCTION_ID)
+            .setRegion(TEST_REGION).addKeyFilter(serializationService.encode("key1"))
+            .addKeyFilter(serializationService.encode("key2")));
+
+    authenticateWithServer();
+    final ClientProtocol.Message responseMessage = writeMessage(message.build());
+
+    FunctionAPI.ExecuteFunctionOnRegionResponse response = getFunctionResponse(responseMessage);
+    assertEquals("result", serializationService.decode(response.getResults(0)));
+
+    final RegionFunctionContext context = (RegionFunctionContext) testFunction.getContext();
+
+    final Set<?> filter = context.getFilter();
+
+    assertEquals(expectedFilter, filter);
+
   }
 
   @Test
@@ -283,10 +305,10 @@ public class ExecutFunctionOnMemberIntegrationTest {
         return Arrays.asList(requiredPermission);
       }
     };
-
     FunctionService.registerFunction(testFunction);
 
     when(securityManager.authenticate(any())).thenReturn(SECURITY_PRINCIPAL);
+
     when(securityManager.authorize(eq(SECURITY_PRINCIPAL), eq(requiredPermission)))
         .thenReturn(false);
 
@@ -297,12 +319,12 @@ public class ExecutFunctionOnMemberIntegrationTest {
     verify(securityManager).authorize(eq(SECURITY_PRINCIPAL), eq(requiredPermission));
   }
 
-  private FunctionAPI.ExecuteFunctionOnMemberResponse getFunctionResponse(
+  private FunctionAPI.ExecuteFunctionOnRegionResponse getFunctionResponse(
       ClientProtocol.Message responseMessage) {
     assertEquals(responseMessage.getResponse().toString(),
-        ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONMEMBERRESPONSE,
+        ClientProtocol.Response.ResponseAPICase.EXECUTEFUNCTIONONREGIONRESPONSE,
         responseMessage.getResponse().getResponseAPICase());
-    return responseMessage.getResponse().getExecuteFunctionOnMemberResponse();
+    return responseMessage.getResponse().getExecuteFunctionOnRegionResponse();
   }
 
   private void authenticateWithServer() throws IOException {
@@ -321,17 +343,17 @@ public class ExecutFunctionOnMemberIntegrationTest {
     authenticateWithServer();
 
     final ClientProtocol.Message request =
-        createRequestMessageBuilder(FunctionAPI.ExecuteFunctionOnMemberRequest.newBuilder()
-            .setFunctionID(TEST_FUNCTION_ID).addMemberName(SERVER_NAME)).build();
+        createRequestMessageBuilder(FunctionAPI.ExecuteFunctionOnRegionRequest.newBuilder()
+            .setFunctionID(TEST_FUNCTION_ID).setRegion(TEST_REGION)).build();
 
     return writeMessage(request);
   }
 
 
   private ClientProtocol.Message.Builder createRequestMessageBuilder(
-      FunctionAPI.ExecuteFunctionOnMemberRequest.Builder functionRequest) {
+      FunctionAPI.ExecuteFunctionOnRegionRequest.Builder functionRequest) {
     return ClientProtocol.Message.newBuilder().setRequest(
-        ClientProtocol.Request.newBuilder().setExecuteFunctionOnMemberRequest(functionRequest));
+        ClientProtocol.Request.newBuilder().setExecuteFunctionOnRegionRequest(functionRequest));
   }
 
   private ClientProtocol.Message writeMessage(ClientProtocol.Message request) throws IOException {
